Move some formTemplate-specific functions to formTemplate.ts
